New changes for statistics - issue #7

+++ src/main/java/com/game/classes/interfaces/model/TopPlayerStats.java
@@ -5,9 +5,9 @@ import java.sql.Date;
 public interface TopPlayerStats {
 	String getUsername();
 
-	int getWinCount();
+	Long getWinCount();
 
-	int getLostCount();
+	Long getLostCount();
 
 	Date getFinishDate();
 }
